Refactor src/world/generation.ts to smaller modules
