+++ src/world/generation.ts
@@ -1,1411 +1,3 @@
-// src/world/generation.ts
-
-import { createIntegratedFireflies, createIntegratedPollen, createIntegratedSparks, registerIntegratedSystem } from '../particles/index.ts';
-import * as THREE from 'three';
-import { getGroundHeight, initCollisionSystem, addCollisionObject, checkPositionValidity } from '../utils/wasm-loader.js';
-import {
-    createSky, createStars, createMoon, createMushroom, createGlowingFlower,
-    createFlower, createSubwooferLotus, createAccordionPalm, createFiberOpticWillow,
-    createFloatingOrb, createSwingableVine, createVineLadder, VineSwing, createPrismRoseBush,
-    createStarflower, createVibratoViolet, createTremoloTulip, createKickDrumGeyser,
-    createRainingCloud, createWaveformWater, initFallingBerries,
-    initGrassSystem, addGrassInstance,
-    createArpeggioFern, createPortamentoPine, createCymbalDandelion, createSnareTrap,
-    createBubbleWillow, createHelixPlant, createBalloonBush,
-    createPanningPad, createSilenceSpirit, createInstrumentShrine, createMelodyMirror,
-    createRetriggerMushroom,
-    createIsland, // Added
-    createCaveEntrance,
-    createTerrainMaterial, // Added
-    createLuminousPlant, luminousPlantBatcher
-} from '../foliage/index.ts';
-import { generateCloudLayer } from '../foliage/procedural-sky.ts';
-import { validateFoliageMaterials, foliageMaterials } from '../foliage/index.ts';
-import { createWisteriaCluster } from '../foliage/wisteria-cluster.ts';
-import { CONFIG } from '../core/config.ts';
-import { generateGroundHeightmap, disposeHeightmap } from './ground-heightmap.ts';
-
-import { registerPhysicsCave } from '../systems/physics/index.js';
-import { initDiscoveryForFoliage } from '../systems/discovery-optimized.ts';
-import { unlockSystem } from '../systems/unlocks.ts';
-import { spawnImpact } from '../foliage/impacts.ts';
-import { makeInteractive } from '../utils/interaction-utils.ts';
-import {
-    animatedFoliage, cpuAnimatedFoliage, obstacles, foliageGroup, foliageMushrooms,
-    foliageClouds, foliageTrampolines, foliagePanningPads, foliageGeysers, foliageTraps, foliagePortamentoPines, vineSwings, foliageVineLadders, worldGroup
-} from './state.ts';
-import mapData from '../../assets/map.json';
-import { globalBackgroundProcessor } from '../utils/background-processor.ts';
-import { showDeferredIndicator, hideDeferredIndicator } from '../ui/index.ts';
-import { recordGenerationChunk } from '../utils/startup-profiler.ts';
-
-// Yields control back to the browser for one event loop turn, allowing the
-// renderer to update the loading screen and preventing "Page Unresponsive" dialogs.
-const yieldControl = (): Promise<void> => new Promise(resolve => setTimeout(resolve, 0));
-
-// Performance constants for async generation
-const obstaclesData: {x: number, y: number, z: number, radius: number}[] = [];
-export const DEFAULT_MAP_CHUNK_SIZE = 100;        // Map entities per chunk (used for progress reporting)
-export const DEFAULT_PROCEDURAL_CHUNK_SIZE = 100; // Procedural extras per chunk
-export const PROCEDURAL_ENTITY_COUNT = 400;       // Number of random procedural items
-/**
- * Maximum ms spent processing map entities before yielding to the event loop.
- * Kept below 16ms so each burst fits within a single animation frame budget,
- * preventing "long task" jank on the main thread.
- */
-const ENTITY_BUDGET_MS = 14;
-const YIELD_ENTITY_BATCH_SIZE = 40;
-const YIELD_LOG_INTERVAL = YIELD_ENTITY_BATCH_SIZE * 5;
-
-function shouldLogYieldProgress(current: number, total: number): boolean {
-    return current === YIELD_ENTITY_BATCH_SIZE || current === total || current % YIELD_LOG_INTERVAL === 0;
-}
-
-// Type definitions for map data
-interface MapEntity {
-    type: string;
-    position: [number, number, number];
-    variant?: string;
-    scale?: number;
-    size?: number | string;
-    note?: string;        // Musical note for mushrooms
-    noteIndex?: number;   // Note index (0-11) for mushrooms
-    hasFace?: boolean;    // Whether mushroom has a face
-}
-
-interface ObstacleData {
-    position: THREE.Vector3;
-    radius: number;
-}
-
-interface WorldObjects {
-    sky: THREE.Object3D;
-    moon: THREE.Object3D;
-    ground: THREE.Mesh;
-}
-
-interface WeatherSystem {
-    registerTree(obj: THREE.Object3D): void;
-    registerShrub(obj: THREE.Object3D): void;
-    registerMushroom(obj: THREE.Object3D): void;
-    registerCave(obj: THREE.Object3D): void;
-}
-
-type WorldProgressCallback = (
-    current: number,
-    total: number,
-    label?: string,
-    entityType?: string
-) => void;
-
-// --- Lake Configuration (Mirrored in Physics.js) ---
-const LAKE_BOUNDS = { minX: -38, maxX: 78, minZ: -28, maxZ: 68 };
-const LAKE_BOTTOM = -2.0;
-
-// --- Lake Island Configuration ---
-// Central island within the lake for musical flora focal point
-const LAKE_ISLAND = {
-    centerX: 20,           // Center X position (middle of lake)
-    centerZ: 20,           // Center Z position (middle of lake)
-    radius: 12,            // Island radius
-    peakHeight: 3.0,       // Height at center above water
-    falloffRadius: 4,      // Smooth blend at edges
-    enabled: true          // Toggle island generation
-};
-
-
-// --- Arpeggio Grove Configuration ---
-const ARPEGGIO_GROVE = {
-    centerX: -60,
-    centerZ: 60,
-    radius: 15,
-    enabled: true
-};
-
-// Helper: Calculate Unified Ground Height (WASM + Visual Lake Modifiers + Island)
-// Matches logic in src/systems/physics.js
-function getUnifiedGroundHeight(x: number, z: number): number {
-    let height = getGroundHeight(x, z);
-
-    // Check if we're in the lake bounds
-    if (x > LAKE_BOUNDS.minX && x < LAKE_BOUNDS.maxX && z > LAKE_BOUNDS.minZ && z < LAKE_BOUNDS.maxZ) {
-        // Calculate distance from lake edges
-        const distX = Math.min(x - LAKE_BOUNDS.minX, LAKE_BOUNDS.maxX - x);
-        const distZ = Math.min(z - LAKE_BOUNDS.minZ, LAKE_BOUNDS.maxZ - z);
-        const distEdge = Math.min(distX, distZ);
-
-        // Check if we're on the island
-        if (LAKE_ISLAND.enabled) {
-            const dx = x - LAKE_ISLAND.centerX;
-            const dz = z - LAKE_ISLAND.centerZ;
-            
-            // ⚡ OPTIMIZATION: Deferred Math.sqrt() by using squared distance for early-out bounds check.
-            const distFromIslandCenterSq = dx * dx + dz * dz;
-            const islandRadiusSq = LAKE_ISLAND.radius * LAKE_ISLAND.radius;
-
-            if (distFromIslandCenterSq < islandRadiusSq) {
-                const distFromIslandCenter = Math.sqrt(distFromIslandCenterSq);
-                // On the island - calculate height above water
-                const normalizedDist = distFromIslandCenter / LAKE_ISLAND.radius;
-                
-                // Smooth falloff using cosine curve for natural hill shape
-                const islandHeight = LAKE_ISLAND.peakHeight * Math.cos(normalizedDist * Math.PI / 2);
-                
-                // Blend at the edge of the island
-                const edgeDist = LAKE_ISLAND.radius - distFromIslandCenter;
-                const edgeBlend = Math.min(1.0, edgeDist / LAKE_ISLAND.falloffRadius);
-                
-                // Island height above water level (water is at ~1.5)
-                const waterLevel = 1.5;
-                const finalIslandHeight = waterLevel + (islandHeight * edgeBlend);
-                
-                // Return island height (don't apply lake depression)
-                return Math.max(height, finalIslandHeight);
-            }
-        }
-
-        // Not on island - apply lake depression
-        const blend = Math.min(1.0, distEdge / 10.0);
-        const targetHeight = THREE.MathUtils.lerp(height, LAKE_BOTTOM, blend);
-
-        if (targetHeight < height) {
-            height = targetHeight;
-        }
-    }
-    return height;
-}
-
-// --- Scene Setup ---
-
-export async function initWorld(scene: THREE.Scene, weatherSystem: WeatherSystem, loadContent: boolean = true): Promise<WorldObjects> {
-    // 0. Pre-flight Check
-    validateFoliageMaterials(foliageMaterials);
-
-    // Sky, stars, moon (fast — no yield needed)
-    const sky = createSky();
-    scene.add(sky);
-
-    const stars = createStars();
-    scene.add(stars);
-
-    const moon = createMoon();
-    moon.position.set(-50, 60, -30); // High up
-    scene.add(moon);
-
-    // Ground - SHRUNK from 2000 to 400 for tighter feel
-    let groundGeo: THREE.PlaneGeometry;
-    let groundMat: THREE.Material;
-
-    // Parse URL parameter for quick toggle
-    const urlParams = new URLSearchParams(window.location.search);
-    const forceGpuTerrain = urlParams.has('gpuTerrain');
-
-    // Yield before the heavy terrain generation so the loading screen can paint.
-    await yieldControl();
-
-    if (CONFIG.terrain?.useGpuHeightmap || forceGpuTerrain) {
-        console.log("[World] Using GPU Heightmap Displacement");
-
-        const resolution = CONFIG.terrain?.heightmapResolution || 256;
-        groundGeo = new THREE.PlaneGeometry(400, 400, resolution, resolution);
-
-        // generateGroundHeightmap is now async and yields internally every 32 rows
-        const startParams = performance.now();
-        const { heightTexture, normalTexture } = await generateGroundHeightmap(400, resolution);
-        console.log(`[World] Generated heightmap in ${(performance.now() - startParams).toFixed(2)}ms`);
-
-        groundMat = createTerrainMaterial(CONFIG.colors.ground, {
-            roughness: 0.9,
-            bumpStrength: 0.15,
-            noiseScale: 20.0
-        }, heightTexture, normalTexture);
-
-    } else {
-        console.log("[World] Using CPU Vertex Displacement");
-
-        groundGeo = new THREE.PlaneGeometry(400, 400, 128, 128);
-        const posAttribute = groundGeo.attributes.position;
-        const vertexCount = posAttribute.count;
-        const cpuYieldEvery = 2000; // yield every ~2k vertices to stay responsive
-
-        for (let i = 0; i < vertexCount; i++) {
-            const x = posAttribute.getX(i);
-            const y = posAttribute.getY(i); // Plane is on XY
-            const zWorld = -y;
-
-            // Use the Unified Height that accounts for the Lake
-            const height = getUnifiedGroundHeight(x, zWorld);
-            posAttribute.setZ(i, height);
-
-            if (i % cpuYieldEvery === cpuYieldEvery - 1) {
-                await yieldControl();
-            }
-        }
-
-        groundGeo.computeVertexNormals();
-
-        // Replaced MeshPhysicalMaterial with Audio-Reactive TSL Material
-        groundMat = createTerrainMaterial(CONFIG.colors.ground, {
-            roughness: 0.9,
-            bumpStrength: 0.15,
-            noiseScale: 20.0
-        });
-    }
-
-    const ground = new THREE.Mesh(groundGeo, groundMat);
-    ground.rotation.x = -Math.PI / 2;
-    ground.receiveShadow = true;
-    scene.add(ground);
-
-    // 2. Update fog colour for compact world.
-    // scene.fogNode (set in init.ts) drives actual WebGPU rendering via TSL rangeFog.
-    // Keep scene.fog as THREE.Fog (not FogExp2) so WeatherSystem's stale reference
-    // stays valid and renderer code never has to read FogExp2.density.
-    const fogColor = new THREE.Color(CONFIG.colors.fog || 0xFFC5D3);
-    if (scene.fog instanceof THREE.Fog) {
-        scene.fog.color.set(fogColor);
-    }
-    scene.background = fogColor;
-
-    // Initialize Vegetation Systems (yield first so browser can breathe)
-    await yieldControl();
-    initGrassSystem(scene, 10000);
-
-    // Use CPU fallback for fireflies during startup. GPU compute init is async but can hang
-    // on systems with partial WebGPU support; the CPU path is safe and fast enough for 150 particles.
-    scene.add(createIntegratedFireflies({ count: 150, areaSize: 100, useCompute: false }));
-
-    // Procedural Cloud Layer (Background)
-    await yieldControl();
-    generateCloudLayer(scene);
-
-    // Melody Lake (Waveform Water)
-    // Lake is at 20, 1.5, 20 with width 120, depth 100
-    const melodyLake = createWaveformWater(120, 100);
-    melodyLake.position.set(20, 1.5, 20); 
-    scene.add(melodyLake);
-
-    // Lake Island
-    const island = createIsland({ radius: 15, height: 2 });
-    island.position.set(-40, 2.5, 40); // Place in the lake
-    island.userData.type = 'lake_island';
-    safeAddFoliage(island, true, 15, weatherSystem);
-
-    // Add Luminous Plants around Lake Island (yield every 30 plants to stay responsive)
-    const luminousCount = CONFIG.luminousPlants.density;
-    await yieldControl();
-    for (let i = 0; i < luminousCount; i++) {
-        const angle = Math.random() * Math.PI * 2;
-        // Gentle radius falloff: more dense near edge (10-25), tapering out to 35
-        // Using a square-root or quadratic distribution helps achieve this
-        const randDist = Math.pow(Math.random(), 2.0); // more values near 0
-        const dist = 10 + randDist * 25; // 10 to 35
-
-        const lx = -40 + Math.cos(angle) * dist;
-        const lz = 40 + Math.sin(angle) * dist;
-        const ly = getUnifiedGroundHeight(lx, lz);
-
-        // Quick biome check to avoid candy cane forest (let's say candy cane is where x > 0)
-        // If x > 0, we'll just skip (assume biome boundary)
-        if (lx > -10) continue;
-
-        // Add a small height bias: prefer elevated ground
-        // Don't spawn directly in water (y < 2.0) and favor y between 2.0 and 5.0
-        if (ly > 2.0 && ly < 8.0) {
-            const plant = createLuminousPlant({ scale: 0.8 + Math.random() * 0.6 });
-            plant.position.set(lx, ly, lz);
-            plant.rotation.y = Math.random() * Math.PI * 2;
-            safeAddFoliage(plant, false, 0, weatherSystem);
-        }
-
-        if (i % 30 === 29) await yieldControl();
-    }
-
-    // Falling Berries
-    await yieldControl();
-    initFallingBerries(scene);
-
-    // Add the luminous plant batcher to the scene
-    scene.add(luminousPlantBatcher.mesh);
-
-    // Add the main world group (containing all generated foliage) to the scene
-    scene.add(worldGroup);
-
-    // Generate Content if requested (triggered by start button in main.ts)
-    if (loadContent) {
-        generateMap(weatherSystem).catch(err => {
-            console.error('[World] Failed to generate map:', err);
-        });
-    }
-
-    return { sky, moon, ground };
-}
-
-
-/**
- * Populates the Arpeggio Grove set piece as defined in the Musical Ecosystem plan.
- * Features a Subwoofer Lotus surrounded by twelve Arpeggio Ferns, with reactive flora.
- * Yields control to the browser between object batches to prevent main-thread blocking.
- */
-async function populateArpeggioGrove(weatherSystem: WeatherSystem): Promise<void> {
-    if (!ARPEGGIO_GROVE.enabled) return;
-
-    console.log("[World] Populating Arpeggio Grove...");
-
-    const { centerX, centerZ, radius } = ARPEGGIO_GROVE;
-
-    // Central feature: Subwoofer Lotus
-    const centralLotus = createSubwooferLotus({ scale: 1.5 });
-    const centralY = getUnifiedGroundHeight(centerX, centerZ);
-    centralLotus.position.set(centerX, centralY, centerZ);
-    safeAddFoliage(centralLotus, false, 0, weatherSystem);
-    await yieldControl();
-
-    // Twelve Arpeggio Ferns ring (yield every 4 ferns)
-    const fernCount = 12;
-    const fernRadius = radius * 0.4;
-    for (let i = 0; i < fernCount; i++) {
-        const angle = (i / fernCount) * Math.PI * 2;
-        const fx = centerX + Math.cos(angle) * fernRadius;
-        const fz = centerZ + Math.sin(angle) * fernRadius;
-        const fy = getUnifiedGroundHeight(fx, fz);
-
-        const fern = createArpeggioFern({ scale: 1.2 + Math.random() * 0.3 });
-        fern.position.set(fx, fy, fz);
-        fern.rotation.y = angle + Math.PI; // Face outward or inward? Let's say outward
-        safeAddFoliage(fern, false, 0, weatherSystem);
-        if (i % 4 === 3) await yieldControl();
-    }
-
-    // Outer ring: Kick Drum Geysers and Vibrato Violets (yield every 4 objects)
-    const outerCount = 8;
-    const outerRadius = radius * 0.8;
-    for (let i = 0; i < outerCount; i++) {
-        const angle = (i / outerCount) * Math.PI * 2 + 0.2;
-        const ox = centerX + Math.cos(angle) * outerRadius;
-        const oz = centerZ + Math.sin(angle) * outerRadius;
-        const oy = getUnifiedGroundHeight(ox, oz);
-
-        if (i % 2 === 0) {
-            const geyser = createKickDrumGeyser({ maxHeight: 5.0 + Math.random() * 2.0 });
-            geyser.position.set(ox, oy, oz);
-            geyser.rotation.y = angle;
-            safeAddFoliage(geyser, false, 1.0, weatherSystem);
-        } else {
-            const violet = createVibratoViolet({ intensity: 1.5 });
-            violet.position.set(ox, oy, oz);
-            violet.rotation.y = Math.random() * Math.PI * 2;
-            safeAddFoliage(violet, false, 0, weatherSystem);
-        }
-        if (i % 4 === 3) await yieldControl();
-    }
-
-    // Glowing flower accents (yield every 5 flowers)
-    const flowerCount = 15;
-    for (let i = 0; i < flowerCount; i++) {
-        const randAngle = Math.random() * Math.PI * 2;
-        const randRadius = Math.random() * radius;
-        const fx = centerX + Math.cos(randAngle) * randRadius;
-        const fz = centerZ + Math.sin(randAngle) * randRadius;
-        const fy = getUnifiedGroundHeight(fx, fz);
-
-        const flower = createGlowingFlower();
-        flower.position.set(fx, fy, fz);
-        flower.rotation.y = Math.random() * Math.PI * 2;
-        safeAddFoliage(flower, false, 0, weatherSystem);
-        if (i % 5 === 4) await yieldControl();
-    }
-
-    console.log(`[World] Arpeggio Grove populated at (${centerX}, ${centerZ})`);
-}
-
-// Ensure the lake island is also present
-
-export function safeAddFoliage(
-    obj: THREE.Object3D,
-    isObstacle: boolean = false,
-    radius: number = 1.0,
-    weatherSystem: WeatherSystem | null = null
-): void {
-    if (animatedFoliage.length > 3000) return; // ⚡ PERFORMANCE: Raised limit from 1000 to 3000 for more musical objects
-    foliageGroup.add(obj);
-    animatedFoliage.push(obj);
-
-    if (!(obj.userData.isBatched ||
-        obj.userData.type === 'mushroom' ||
-        obj.userData.type === 'lanternFlower' ||
-        obj.userData.type === 'arpeggio_fern' ||
-        obj.userData.type === 'portamento_pine' ||
-        obj.userData.type === 'prismRoseBush' ||
-        obj.userData.isFlower)) {
-        cpuAnimatedFoliage.push(obj);
-    }
-
-    // Add to JS obstacles (legacy/backup)
-    if (isObstacle) {
-        // obstacles.push({ position: obj.position.clone(), radius }); // Replaced by obstaclesData
-
-        // ⚡ OPTIMIZATION: Add to WASM Spatial Grid for O(1) validity checks later in batch
-        // Type 5 = Generic Obstacle (Radius Check Only)
-        obstaclesData.push({ x: obj.position.x, y: obj.position.y, z: obj.position.z, radius });
-    }
-
-    // Optimization
-    if (obj.userData.type === 'mushroom') foliageMushrooms.push(obj);
-    if (obj.userData.type === 'cloud') foliageClouds.push(obj);
-    if (obj.userData.isTrampoline) foliageTrampolines.push(obj);
-    if (obj.userData.type === 'panningPad') foliagePanningPads.push(obj);
-    if (obj.userData.type === 'geyser') foliageGeysers.push(obj);
-    if (obj.userData.type === 'trap') {
-        foliageTraps.push(obj);
-        console.log('[World] Registered Snare Trap. Total:', foliageTraps.length);
-    }
-    if (obj.userData.type === 'tree' && obj.userData.animationType === 'batchedPortamento') {
-        foliagePortamentoPines.push(obj);
-    }
-    if (obj.userData.type === 'vine_ladder') foliageVineLadders.push(obj);
-
-    // Invoke deferred placement logic (e.g. for batching)
-    if (obj.userData.onPlacement) {
-        obj.userData.onPlacement();
-    }
-
-    // Register with weather system
-    if (weatherSystem) {
-        if (obj.userData.type === 'tree') {
-            weatherSystem.registerTree(obj);
-        } else if (obj.userData.type === 'shrub') {
-            weatherSystem.registerShrub(obj);
-        } else if (obj.userData.type === 'mushroom') {
-            weatherSystem.registerMushroom(obj);
-        } else if (obj.userData.type === 'cave') {
-            weatherSystem.registerCave(obj);
-            registerPhysicsCave(obj);
-        }
-    }
-}
-
-// --- HELPER: Position Validation ---
-function isPositionValid(x: number, z: number, radius: number): boolean {
-    const distFromCenterSq = x * x + z * z;
-    if (distFromCenterSq < 15 * 15) return false;
-
-    // ⚡ PERFORMANCE: Use WASM Spatial Grid for O(1) check instead of O(N) loop
-    const isValidWasm = checkPositionValidity(x, z, radius);
-    if (isValidWasm === 1) return false; // 1 = Collision
-
-    /* Legacy O(N) Loop - Kept for reference
-    for (const obs of obstacles) {
-        const dx = x - obs.position.x;
-        const dz = z - obs.position.z;
-        const distSq = dx * dx + dz * dz;
-        const minDistance = obs.radius + radius + 1.5;
-        if (distSq < minDistance * minDistance) return false;
-    }
-    */
-    
-    // 3. Lake Avoidance for PROCEDURAL content
-    // We specifically prevent random generation in the lake so we don't drown bushes.
-    // However, map.json entities or explicitly placed objects (like the Cave) are allowed.
-    if (x > -40 && x < 80 && z > -30 && z < 70) {
-        return false;
-    }
-
-    return true;
-}
-
-
-// --- MAP GENERATION ---
-
-/**
- * Async map generation that processes entities in chunks to prevent UI freeze
- * @param weatherSystem The weather system to register entities with
- * @param chunkSize Number of entities to process per frame (default: 100)
- * @param onProgress Optional callback for progress updates
- */
-
-
-/**
- * Determine if an entity is critical for initial load (collision, physics, interaction)
- */
-export function isCriticalEntity(item: MapEntity | { type: string, isObstacle?: boolean }): boolean {
-    const criticalTypes = [
-        'mushroom', // Often giant / bouncy
-        'tree',     // usually has collision
-        'arpeggio_fern',
-        'portamento_pine',
-        'snare_trap',
-        'geyser',
-        'trap',
-        'panningPad',
-        'cloud',    // can be Walkable
-        'vine_ladder',
-        'instrumentShrine',
-        'waterfall'
-    ];
-
-    if (criticalTypes.includes(item.type)) return true;
-
-    // Explicit overrides
-    if ((item as any).critical === true) return true;
-    if ((item as any).isObstacle === true) return true;
-    if (item.type === 'cave') return true;
-
-    return false;
-}
-
-export async function generateMap(
-    weatherSystem: WeatherSystem, 
-    chunkSize: number = DEFAULT_MAP_CHUNK_SIZE,
-    onProgress?: WorldProgressCallback
-): Promise<void> {
-    performance.mark('candy:map-generation-start');
-    console.log(`[World] Loading map with ${mapData.entities.length} entities...`);
-
-    // Reset WASM Collision System for Generation Phase
-    initCollisionSystem();
-
-    const entities = mapData.entities as any as MapEntity[];
-    
-    // 1. Filter into Critical vs Deferred
-    const criticalEntities: MapEntity[] = [];
-    const deferredEntities: MapEntity[] = [];
-
-    let scannedEntities = 0;
-    for (const item of entities) {
-        if (isCriticalEntity(item)) {
-            criticalEntities.push(item);
-        } else {
-            deferredEntities.push(item);
-        }
-
-        scannedEntities++;
-        if (scannedEntities % YIELD_ENTITY_BATCH_SIZE === 0 && scannedEntities < entities.length) {
-            if (shouldLogYieldProgress(scannedEntities, entities.length)) {
-                console.log(`[World] Yielding during entity scan at ${scannedEntities}/${entities.length} (last type: ${item.type})`);
-            }
-            await yieldControl();
-        }
-    }
-
-    console.log(`[World] Filtered map: ${criticalEntities.length} critical, ${deferredEntities.length} deferred.`);
-
-    const criticalTotal = criticalEntities.length;
-    // For progress, we only track critical + Arpeggio Grove (treated as 1 chunk) during the blocking phase
-    // Procedural extras are now split as well. We'll approximate.
-    const globalTotal = criticalTotal + 1; // +1 for Arpeggio Grove
-
-    // 2. Process Critical Map Entities with a per-entity time budget.
-    // Instead of fixed-size chunks (which could still run long if entities are heavy),
-    // we check elapsed time after each entity and yield as soon as ENTITY_BUDGET_MS
-    // is exceeded. This keeps every task well under 100 ms.
-    let i = 0;
-    while (i < criticalTotal) {
-        const chunkStart = performance.now();
-        let processed = 0;
-        let lastEntityType = 'entity';
-
-        while (i + processed < criticalTotal) {
-            const idx = i + processed;
-            const entity = criticalEntities[idx];
-            lastEntityType = entity.type;
-            processMapEntity(entity, weatherSystem);
-            processed++;
-
-            // Granular progress update every 50 entities with detailed text
-            if ((i + processed) % 50 === 0) {
-                const percentage = Math.floor(((i + processed) / criticalTotal) * 100);
-                updateProgress('map-generation', percentage, `Spawning flora: ${i + processed}/${criticalTotal}`);
-            }
-
-            // Yield as soon as we've spent our per-chunk budget.
-            if (processed >= YIELD_ENTITY_BATCH_SIZE || performance.now() - chunkStart >= ENTITY_BUDGET_MS) {
-                break;
-            }
-        }
-
-        i += processed;
-
-        if (onProgress) {
-            const current = Math.min(i, criticalTotal);
-            onProgress(
-                current,
-                globalTotal,
-                `[World] Populating world ${current}/${criticalTotal}`,
-                lastEntityType
-            );
-        }
-
-        // Record this chunk for the startup profiler.
-        recordGenerationChunk();
-
-        // Yield control back to the browser.
-        if (i < criticalTotal) {
-            if (shouldLogYieldProgress(i, criticalTotal)) {
-                console.log(`[World] Yielding after ${i}/${criticalTotal} critical entities (last type: ${lastEntityType})`);
-            }
-            await yieldControl();
-        }
-    }
-
-    // --- Spawn The Cave (Critical) ---
-    const cave = createCaveEntrance({ scale: 2.0 });
-    const caveX = 25;
-    const caveZ = 25;
-    const caveY = getUnifiedGroundHeight(caveX, caveZ);
-    cave.position.set(caveX, caveY, caveZ);
-    cave.lookAt(0, caveY, 0);
-    safeAddFoliage(cave, false, 0, weatherSystem);
-    console.log("[World] Cave spawned at ", caveX, caveZ, " Height:", caveY);
-
-    if (cave.userData.gatePosition) {
-        const waterfallProxy = new THREE.Object3D();
-        cave.updateMatrixWorld(true);
-        waterfallProxy.position.copy(cave.userData.gatePosition).applyMatrix4(cave.matrixWorld);
-        waterfallProxy.userData.type = 'waterfall';
-        animatedFoliage.push(waterfallProxy as any);
-    }
-
-    // --- Populate Arpeggio Grove Set Piece (Critical) ---
-    await populateArpeggioGrove(weatherSystem);
-    if (onProgress) {
-        onProgress(globalTotal, globalTotal, '[World] Critical full world population complete', 'arpeggio_grove');
-    }
-    
-    // --- Initialize Discovery System with Spatial Grid (Critical) ---
-    // OPTIMIZATION: O(1) spatial lookups instead of O(N) distance checks
-    // We do this NOW before deferring the rest, so grids are static and complete for interactive items
-    initDiscoveryForFoliage(animatedFoliage);
-    
-    // 3. Queue Deferred Map Entities
-    console.log(`[World] Queueing ${deferredEntities.length} deferred map entities for background processing...`);
-    let queuedDeferred = 0;
-    for (const item of deferredEntities) {
-        globalBackgroundProcessor.enqueue({
-            id: `map_deferred_${item.type}_${Math.random()}`,
-            execute: () => processMapEntity(item, weatherSystem)
-        });
-
-        queuedDeferred++;
-        if (queuedDeferred % YIELD_ENTITY_BATCH_SIZE === 0 && queuedDeferred < deferredEntities.length) {
-            if (shouldLogYieldProgress(queuedDeferred, deferredEntities.length)) {
-                console.log(`[World] Yielding while queueing deferred entities at ${queuedDeferred}/${deferredEntities.length} (last type: ${item.type})`);
-            }
-            await yieldControl();
-        }
-    }
-
-    // --- Populate Procedural Extras (Split into Critical/Deferred inside) ---
-    await populateProceduralExtras(weatherSystem, chunkSize);
-
-    performance.mark('candy:map-generation-end');
-    try {
-        performance.measure('candy:Map Generation', 'candy:map-generation-start', 'candy:map-generation-end');
-    } catch (_e) { /* ignore if marks were cleared */ }
-
-    console.log("[World] Critical map generation complete! Deferred tasks queued.");
-}
-
-export async function generateCoreWorld(
-    weatherSystem: WeatherSystem,
-    onProgress?: WorldProgressCallback
-): Promise<void> {
-    console.log('[World] Core Only mode: generating lightweight candy landscape');
-    initCollisionSystem();
-
-    const areaSize = 120;
-    const maxAttempts = 20;
-    const getRandomGroundPosition = (radius: number) => {
-        for (let attempt = 0; attempt < maxAttempts; attempt++) {
-            const x = (Math.random() - 0.5) * areaSize;
-            const z = (Math.random() - 0.5) * areaSize;
-            if (!isPositionValid(x, z, radius)) continue;
-            return { x, z, y: getUnifiedGroundHeight(x, z) };
-        }
-        return null;
-    };
-
-    if (onProgress) onProgress(0, 4, '[World] Generating core world');
-
-    // Basic candy trees — yield every ENTITY_BUDGET_MS to avoid blocking the main thread.
-    // Tree geometry creation can take 10–30 ms each; without yielding 18 trees back-to-back
-    // would stall the browser for up to 540 ms and trigger "Page Unresponsive".
-    const treeFactories = [
-        () => createBubbleWillow(),
-        () => createBalloonBush(),
-        () => createHelixPlant(),
-        () => createPortamentoPine({ height: 4.5 }),
-    ];
-    let chunkStart = performance.now();
-    for (let i = 0; i < 18; i++) {
-        const factory = treeFactories[i % treeFactories.length];
-        const pos = getRandomGroundPosition(1.5);
-        if (pos) {
-            const obj = factory();
-            obj.position.set(pos.x, pos.y, pos.z);
-            obj.rotation.y = Math.random() * Math.PI * 2;
-            safeAddFoliage(obj, true, 1.5, weatherSystem);
-        }
-        if (performance.now() - chunkStart >= ENTITY_BUDGET_MS) {
-            await yieldControl();
-            chunkStart = performance.now();
-        }
-    }
-    if (onProgress) onProgress(1, 4, '[World] Core trees ready', 'tree');
-
-    // Mushrooms and ground accents — same time-based yield approach.
-    chunkStart = performance.now();
-    for (let i = 0; i < 24; i++) {
-        const pos = getRandomGroundPosition(0.5);
-        if (pos) {
-            const obj = createMushroom({ size: 'regular', scale: 0.8 + Math.random() * 0.5, hasFace: true, isBouncy: true });
-            obj.position.set(pos.x, pos.y, pos.z);
-            obj.rotation.y = Math.random() * Math.PI * 2;
-            safeAddFoliage(obj, true, 0.5, weatherSystem);
-        }
-        if (performance.now() - chunkStart >= ENTITY_BUDGET_MS) {
-            await yieldControl();
-            chunkStart = performance.now();
-        }
-    }
-    if (onProgress) onProgress(2, 4, '[World] Core mushrooms ready', 'mushroom');
-
-    // Clouds above the terrain.
-    chunkStart = performance.now();
-    for (let i = 0; i < 12; i++) {
-        const pos = getRandomGroundPosition(0.8);
-        if (!pos) continue;
-        const height = 10 + Math.random() * 18;
-        const cloud = createRainingCloud({ size: 1.0 + Math.random() * 0.8 });
-        cloud.position.set(pos.x, height, pos.z);
-        cloud.userData.tier = 1;
-        cloud.userData.isWalkable = true;
-        safeAddFoliage(cloud, false, 0.8, weatherSystem);
-        if (performance.now() - chunkStart >= ENTITY_BUDGET_MS) {
-            await yieldControl();
-            chunkStart = performance.now();
-        }
-    }
-    if (onProgress) onProgress(3, 4, '[World] Core clouds ready', 'cloud');
-
-    // Low flowers and luminous accents (lightweight — single yield at end is sufficient).
-    for (let i = 0; i < 16; i++) {
-        const factory = Math.random() < 0.5 ? () => createFlower() : () => createGlowingFlower();
-        const pos = getRandomGroundPosition(0.4);
-        if (pos) {
-            const obj = factory();
-            obj.position.set(pos.x, pos.y, pos.z);
-            obj.rotation.y = Math.random() * Math.PI * 2;
-            safeAddFoliage(obj, false, 0.4, weatherSystem);
-        }
-    }
-    await yieldControl();
-
-    // Lake island accents.
-    const islandItems = [
-        () => createGlowingFlower(),
-        () => createFlower(),
-    ];
-    for (let i = 0; i < 8; i++) {
-        const pos = getRandomGroundPosition(0.4);
-        if (!pos) continue;
-        const factory = islandItems[i % islandItems.length];
-        const obj = factory();
-        obj.position.set(pos.x, pos.y, pos.z);
-        obj.rotation.y = Math.random() * Math.PI * 2;
-        safeAddFoliage(obj, false, 0.4, weatherSystem);
-    }
-
-    initDiscoveryForFoliage(animatedFoliage);
-    if (onProgress) onProgress(4, 4, '[World] Core world population complete', 'flower');
-    console.log(`[World] Core Only world generation complete. Spawned ${animatedFoliage.length} objects.`);
-}
-
-export type WorldMode = 'CORE' | 'FULL';
-
-export async function populateWorld(
-    scene: THREE.Scene,
-    weatherSystem: WeatherSystem,
-    mode: WorldMode = 'CORE',
-    onProgress?: WorldProgressCallback
-): Promise<WorldMode> {
-    console.log(`[World] Starting populateWorld() in ${mode} mode`);
-
-    if (mode === 'CORE') {
-        console.log('%c[World] CORE Mode active — spawning minimal classic candy set', 'color:#ff9ecd');
-        console.log('[World] Core mode skips: map entities, arpeggio grove, procedural extras, WASM physics upload');
-        await generateCoreWorld(weatherSystem, onProgress);
-        console.log('[World] Core mode ready. Heavy foliage systems skipped.');
-        console.log('[World] populateWorld() complete in CORE mode');
-        return 'CORE';
-    }
-
-    console.log('%c[World] FULL Mode — attempting complete musical ecosystem', 'color:#7dd3fc');
-    console.log(`[World] Full mode: ${(mapData as any).entities?.length ?? 0} map entities + ${PROCEDURAL_ENTITY_COUNT} procedural extras to process`);
-    try {
-        await generateMap(weatherSystem, DEFAULT_MAP_CHUNK_SIZE, onProgress);
-        console.log('[World] Full mode population complete.');
-        console.log('[World] populateWorld() complete in FULL mode');
-        return 'FULL';
-    } catch (error) {
-        console.error('[World] Full population failed. Falling back from FULL to CORE.', error);
-        await generateCoreWorld(weatherSystem, onProgress);
-        console.log('[World] populateWorld() recovered in CORE mode after FULL failure');
-        return 'CORE';
-    }
-}
-
-/**
- * Process a single map entity (extracted from forEach loop for chunking)
- */
-function processMapEntity(item: MapEntity, weatherSystem: WeatherSystem): void {
-    const [x, yInput, z] = item.position;
-    // USE UNIFIED HEIGHT for placement
-    const groundY = getUnifiedGroundHeight(x, z);
-    let y = groundY;
-    if (item.type === 'cloud') y = yInput;
-
-    try {
-        let obj: THREE.Object3D | null = null;
-        let isObstacle = false;
-        let radius = 0.5;
-
-        // --- Basic Types ---
-        if (item.type === 'mushroom') {
-                const isGiant = item.variant === 'giant';
-                const scale = item.scale || 1.0;
-                const hasFace = item.hasFace !== undefined ? item.hasFace : (isGiant || Math.random() < 0.1);
-                const isBouncy = isGiant || hasFace;
-
-                obj = createMushroom({
-                    size: isGiant ? 'giant' : 'regular',
-                    scale,
-                    hasFace,
-                    isBouncy,
-                    note: item.note,
-                    noteIndex: item.noteIndex
-                });
-                isObstacle = true;
-                radius = isGiant ? 2.0 : 0.5;
-            }
-        else if (item.type === 'flower') {
-            const isGlowing = item.variant === 'glowing';
-            obj = isGlowing ? createGlowingFlower() : createFlower();
-        }
-        else if (item.type === 'cloud') {
-            const cloudTier = (item as any).tier || 1;
-            obj = createRainingCloud({ size: item.size as number || 1.5 });
-            obj.userData.tier = cloudTier;
-            obj.userData.isWalkable = cloudTier === 1;
-        }
-        else if (item.type === 'grass') {
-            addGrassInstance(x, y, z);
-            return;
-        }
-        // ... (Other types elided for brevity, same logic follows) ...
-        else if (item.type === 'subwoofer_lotus') {
-            obj = createSubwooferLotus({ scale: item.scale || 1.0 });
-        }
-        else if (item.type === 'accordion_palm') {
-            obj = createAccordionPalm({ color: 0xFFD700 });
-            isObstacle = true;
-        }
-        else if (item.type === 'fiber_optic_willow') {
-            obj = createFiberOpticWillow();
-            isObstacle = true;
-        }
-        else if (item.type === 'floating_orb') {
-            obj = createFloatingOrb({ size: 0.5 });
-            y += 1.5;
-        }
-        else if (item.type === 'swingable_vine') {
-            obj = createSwingableVine({ length: 8 });
-            y += 8;
-            if (vineSwings) vineSwings.push(new VineSwing(obj, 8));
-        }
-        else if (item.type === 'prism_rose_bush') {
-            obj = createPrismRoseBush();
-            isObstacle = true;
-        }
-        else if (item.type === 'starflower') {
-            obj = createStarflower();
-        }
-        else if (item.type === 'vibrato_violet') {
-            obj = createVibratoViolet();
-        }
-        else if (item.type === 'tremolo_tulip') {
-            obj = createTremoloTulip();
-        }
-        else if (item.type === 'kick_drum_geyser') {
-            obj = createKickDrumGeyser();
-        }
-        // Musical Flora
-        else if (item.type === 'arpeggio_fern') {
-            obj = createArpeggioFern({ scale: item.scale || 1.0 });
-        }
-        else if (item.type === 'portamento_pine') {
-            obj = createPortamentoPine({ height: 4.0 });
-            isObstacle = true;
-        }
-        else if (item.type === 'cymbal_dandelion') {
-            obj = createCymbalDandelion();
-        }
-        else if (item.type === 'snare_trap') {
-            obj = createSnareTrap();
-        }
-        else if (item.type === 'retrigger_mushroom') {
-            obj = createRetriggerMushroom({ scale: item.scale || 1.0 });
-        }
-        else if (item.type === 'panning_pad') {
-            const panBias = x < 0 ? -1 : 1;
-            obj = createPanningPad({ radius: item.scale || 1.0, panBias: panBias });
-            if (y < 2) y = 1.0;
-        }
-        // Spirits
-        else if (item.type === 'silence_spirit') {
-            obj = createSilenceSpirit({ scale: item.scale || 1.0 });
-        }
-        // Instrument Shrines
-        else if (item.type === 'instrument_shrine') {
-            const id = parseInt(item.variant || '0', 10);
-            obj = createInstrumentShrine({ instrumentID: id, scale: item.scale || 1.0 });
-            isObstacle = true;
-            radius = 1.0;
-        }
-        // Trees
-        else if (item.type === 'bubble_willow') {
-            obj = createBubbleWillow();
-            isObstacle = true;
-        }
-        else if (item.type === 'helix_plant') {
-            obj = createHelixPlant();
-        }
-        else if (item.type === 'balloon_bush') {
-            obj = createBalloonBush();
-        }
-        else if (item.type === 'wisteria_cluster') {
-            obj = createWisteriaCluster();
-            y += 4;
-        }
-
-        // --- Spawning ---
-        if (obj) {
-            obj.position.set(x, y, z);
-            obj.rotation.y = Math.random() * Math.PI * 2;
-            if (item.scale && item.type !== 'mushroom' && item.type !== 'flower') {
-                obj.scale.setScalar(item.scale);
-            }
-            safeAddFoliage(obj, isObstacle, radius, weatherSystem);
-        }
-
-    } catch (e) {
-        console.warn(`[World] Failed to spawn ${item.type} at ${x},${z}`, e);
-    }
-}
-
-/**
- * Populates the Lake Island with a curated selection of musical flora.
- * The island serves as a focal point for audio-reactive elements.
- */
-function populateLakeIsland(weatherSystem: WeatherSystem): void {
-    if (!LAKE_ISLAND.enabled) return;
-    
-    console.log("[World] Populating Lake Island with musical flora...");
-    
-    const { centerX, centerZ, radius, peakHeight } = LAKE_ISLAND;
-    
-    // Central feature: Large Retrigger Mushroom
-    const centralMushroom = createRetriggerMushroom({ 
-        scale: 1.5, 
-        retriggerSpeed: 4,
-        color: 0x00FFFF 
-    });
-    const centralY = getUnifiedGroundHeight(centerX, centerZ);
-    centralMushroom.position.set(centerX, centralY, centerZ);
-    makeInteractive(centralMushroom);
-    centralMushroom.userData.interactionText = "Harvest Lake Core";
-    centralMushroom.userData.onInteract = () => {
-        unlockSystem.harvest('lake_core', 1, 'Lake Core');
-        spawnImpact(centralMushroom.position, 'spore', 0x00FFFF);
-        centralMushroom.userData.interactionText = "Harvested";
-        centralMushroom.userData.onInteract = undefined;
-    };
-    safeAddFoliage(centralMushroom, false, 0, weatherSystem);
-    
-    // Ring of Kick Drum Geysers around the perimeter
-    const geyserCount = 6;
-    const geyserRadius = radius * 0.7;
-    for (let i = 0; i < geyserCount; i++) {
-        const angle = (i / geyserCount) * Math.PI * 2;
-        const gx = centerX + Math.cos(angle) * geyserRadius;
-        const gz = centerZ + Math.sin(angle) * geyserRadius;
-        const gy = getUnifiedGroundHeight(gx, gz);
-        
-        const geyser = createKickDrumGeyser({ maxHeight: 4.0 + Math.random() * 2.0 });
-        geyser.position.set(gx, gy, gz);
-        geyser.rotation.y = angle + Math.PI; // Face outward
-        safeAddFoliage(geyser, false, 1.0, weatherSystem);
-    }
-    
-    // Inner ring: Alternating Vibrato Violets and Tremolo Tulips
-    const flowerCount = 8;
-    const flowerRadius = radius * 0.45;
-    for (let i = 0; i < flowerCount; i++) {
-        const angle = (i / flowerCount) * Math.PI * 2 + 0.3; // Offset from geysers
-        const fx = centerX + Math.cos(angle) * flowerRadius;
-        const fz = centerZ + Math.sin(angle) * flowerRadius;
-        const fy = getUnifiedGroundHeight(fx, fz);
-        
-        const flower = i % 2 === 0 
-            ? createVibratoViolet({ intensity: 1.2 })
-            : createTremoloTulip({ size: 1.2 });
-        flower.position.set(fx, fy, fz);
-        flower.rotation.y = Math.random() * Math.PI * 2;
-        safeAddFoliage(flower, false, 0, weatherSystem);
-    }
-    
-    // Scattered Arpeggio Ferns
-    const fernCount = 5;
-    for (let i = 0; i < fernCount; i++) {
-        // Random position within island
-        const randAngle = Math.random() * Math.PI * 2;
-        const randRadius = Math.random() * (radius * 0.6);
-        const fx = centerX + Math.cos(randAngle) * randRadius;
-        const fz = centerZ + Math.sin(randAngle) * randRadius;
-        const fy = getUnifiedGroundHeight(fx, fz);
-        
-        const fern = createArpeggioFern({ scale: 0.8 + Math.random() * 0.4 });
-        fern.position.set(fx, fy, fz);
-        fern.rotation.y = Math.random() * Math.PI * 2;
-        safeAddFoliage(fern, false, 0, weatherSystem);
-    }
-    
-    // Edge decorations: Cymbal Dandelions
-    const dandelionCount = 10;
-    for (let i = 0; i < dandelionCount; i++) {
-        const angle = (i / dandelionCount) * Math.PI * 2 + Math.random() * 0.2;
-        const edgeOffset = radius * 0.85 + Math.random() * (radius * 0.1);
-        const dx = centerX + Math.cos(angle) * edgeOffset;
-        const dz = centerZ + Math.sin(angle) * edgeOffset;
-        const dy = getUnifiedGroundHeight(dx, dz);
-        
-        // Only place if we're still above water
-        if (dy > 1.6) {
-            const dandelion = createCymbalDandelion({ scale: 0.7 + Math.random() * 0.3 });
-            dandelion.position.set(dx, dy, dz);
-            dandelion.rotation.y = Math.random() * Math.PI * 2;
-            safeAddFoliage(dandelion, false, 0, weatherSystem);
-        }
-    }
-    
-    // Corner accent: Snare Traps near the edges
-    const trapCount = 3;
-    for (let i = 0; i < trapCount; i++) {
-        const angle = (i / trapCount) * Math.PI * 2 + Math.PI / 6;
-        const tx = centerX + Math.cos(angle) * (radius * 0.55);
-        const tz = centerZ + Math.sin(angle) * (radius * 0.55);
-        const ty = getUnifiedGroundHeight(tx, tz);
-        
-        const trap = createSnareTrap({ scale: 0.9 });
-        trap.position.set(tx, ty, tz);
-        trap.rotation.y = angle;
-        safeAddFoliage(trap, true, 0.8, weatherSystem);
-    }
-
-    // ⚡ JUICE: Neon Pollen Cloud
-    // Audio-reactive magic dust covering the island
-    const pollen = createIntegratedPollen({ count: 100, areaSize: 25, center: new THREE.Vector3(centerX, 5, centerZ), useCompute: true });
-    safeAddFoliage(pollen, false, 0, null);
-    if ((pollen as any).userData?.computeParticleSystem) {
-        registerIntegratedSystem('pollen_island', pollen, (pollen as any).userData.computeParticleSystem);
-    }
-
-    // ⚡ JUICE: Environmental Sparks around the Core
-    const ambientSparks = createIntegratedSparks({ count: 100, areaSize: 15, center: new THREE.Vector3(centerX, 2, centerZ), useCompute: true });
-    safeAddFoliage(ambientSparks, false, 0, null);
-    if ((ambientSparks as any).userData?.computeParticleSystem) {
-        registerIntegratedSystem('sparks_island', ambientSparks, (ambientSparks as any).userData.computeParticleSystem);
-    }
-
-    const ambientIslandSparks = createIntegratedSparks({ count: 100, areaSize: 15, center: new THREE.Vector3(centerX, 2, centerZ), useCompute: true });
-    safeAddFoliage(ambientIslandSparks, false, 0, null);
-    if ((ambientIslandSparks as any).userData?.computeParticleSystem) {
-        registerIntegratedSystem('sparks_island', ambientIslandSparks, (ambientIslandSparks as any).userData.computeParticleSystem);
-    }
-
-    // ⚡ JUICE: Environmental Sparks
-    // Add ambient sparks to the world
-    const sparksAmbient = createIntegratedSparks({ count: 100, areaSize: 50, center: new THREE.Vector3(centerX, 10, centerZ), useCompute: true });
-    safeAddFoliage(sparksAmbient, false, 0, null);
-    const globalSparks = createIntegratedSparks({ count: 100, areaSize: 50, center: new THREE.Vector3(centerX, 10, centerZ), useCompute: true });
-    safeAddFoliage(globalSparks, false, 0, null);
-
-    console.log(`[World] Lake Island populated with musical flora at (${centerX}, ${centerZ})`);
-}
-
-async function populateProceduralExtras(
-    weatherSystem: WeatherSystem,
-    chunkSize: number = DEFAULT_PROCEDURAL_CHUNK_SIZE
-): Promise<void> {
-    console.log("[World] Populating procedural extras (Critical + Deferred)...");
-    const extrasCount = PROCEDURAL_ENTITY_COUNT;
-    const range = 150;
-
-    // We no longer block the main thread for non-critical procedural objects.
-    // Instead, we immediately calculate their positions and types.
-    // If they are critical, we spawn them now.
-    // If deferred, we queue them in the BackgroundProcessor.
-
-    let criticalCount = 0;
-    let deferredCount = 0;
-    // Track elapsed time so we can yield after each critical spawn that hits the budget.
-    let chunkStart = performance.now();
-
-    for (let i = 0; i < extrasCount; i++) {
-        let x = 0, z = 0, y = 0;
-        let attempts = 0;
-        let validPosition = false;
-
-        while (attempts < 10) {
-            x = (Math.random() - 0.5) * range;
-            z = (Math.random() - 0.5) * range;
-            if (isPositionValid(x, z, 1.5)) {
-                validPosition = true;
-                break;
-            }
-            attempts++;
-        }
-
-        if (!validPosition) continue;
-        const groundY = getUnifiedGroundHeight(x, z);
-
-        // Determine object type and criticality
-        const rand = Math.random();
-
-        // Define a closure to spawn this specific extra
-        const spawnExtra = () => {
-            let obj: THREE.Object3D | null = null;
-            let isObstacle = false;
-            let radius = 0.5;
-            let currentY = groundY;
-
-            try {
-                if (rand < 0.3) {
-                     obj = Math.random() < 0.5 ? createFlower() : createGlowingFlower();
-                     obj.position.set(x, currentY, z);
-                }
-                else if (rand < 0.45) {
-                     obj = createMushroom({
-                         size: 'regular',
-                         scale: 0.8 + Math.random() * 0.5,
-                         hasFace: true,
-                         isBouncy: true
-                     });
-                     obj.position.set(x, currentY, z);
-                     isObstacle = true;
-                }
-                else if (rand < 0.55) {
-                     const treeType = Math.random();
-                     if (treeType < 0.33) obj = createBubbleWillow();
-                     else if (treeType < 0.66) obj = createBalloonBush();
-                     else obj = createHelixPlant();
-
-                     obj.position.set(x, currentY, z);
-                     isObstacle = true;
-                     radius = 1.5;
-                }
-                else if (rand < 0.75) {
-                     const type = Math.random();
-                     if (type < 0.15) {
-                         obj = createArpeggioFern({ scale: 1.0 + Math.random() * 0.5 });
-                     } else if (type < 0.28) {
-                         obj = createKickDrumGeyser({ maxHeight: 5.0 + Math.random() * 3.0 });
-                         radius = 1.0;
-                     } else if (type < 0.40) {
-                         obj = createSnareTrap({ scale: 0.8 + Math.random() * 0.4 });
-                         isObstacle = true;
-                         radius = 0.8;
-                     } else if (type < 0.50) {
-                         obj = createRetriggerMushroom({ scale: 0.8 + Math.random() * 0.4, retriggerSpeed: 2 + Math.floor(Math.random() * 6) });
-                     } else if (type < 0.60) {
-                         obj = createPortamentoPine({ height: 4.0 + Math.random() * 2.0 });
-                         isObstacle = true;
-                         radius = 0.5;
-                     } else if (type < 0.75) {
-                         obj = createTremoloTulip({ size: 1.0 + Math.random() * 0.5 });
-                     } else if (type < 0.85) {
-                         obj = createCymbalDandelion({ scale: 0.8 + Math.random() * 0.4 });
-                     } else {
-                         const panBias = x < 0 ? -1 : 1;
-                         obj = createPanningPad({ radius: 1.2 + Math.random(), panBias });
-                         currentY = groundY + 0.5;
-                         obj.position.y = currentY;
-                     }
-                     if (obj) obj.position.set(x, currentY, z);
-                }
-                 else if (rand < 0.90) {
-                     // Vertical Ecosystem: Tiered Clouds
-                     const tierRoll = Math.random();
-                     if (tierRoll < 0.35) {
-                         currentY = 35 + Math.random() * 20;
-                         obj = createRainingCloud({ size: 1.5 + Math.random() * 0.8 });
-                         obj.userData.tier = 1;
-                         obj.userData.isWalkable = true;
-                         if (Math.random() < 0.3) {
-                             const ladderLength = currentY - groundY;
-                             if (ladderLength > 5) {
-                                 const ladder = createVineLadder({ length: ladderLength });
-                                 ladder.position.set(x, currentY, z);
-                                 safeAddFoliage(ladder, false, 0, weatherSystem);
-                             }
-                         }
-                     } else {
-                         currentY = 12 + Math.random() * 16;
-                         obj = createRainingCloud({ size: 0.8 + Math.random() * 0.6 });
-                         obj.userData.tier = 2;
-                         obj.userData.isWalkable = false;
-                     }
-                     obj.position.set(x, currentY, z);
-                 }
-                 else if (rand < 0.95) {
-                     obj = createSilenceSpirit();
-                     obj.position.set(x, currentY, z);
-                 }
-                 else if (rand < 0.97) {
-                     obj = createMelodyMirror({ scale: 2.0 });
-                     obj.position.set(x, groundY + 15 + Math.random() * 10, z);
-                 }
-             else {
-                 const id = Math.floor(Math.random() * 16);
-                 obj = createInstrumentShrine({ instrumentID: id });
-                 obj.position.set(x, currentY, z);
-                 isObstacle = true;
-            }
-
-            if (obj) {
-                obj.rotation.y = Math.random() * Math.PI * 2;
-                safeAddFoliage(obj, isObstacle, radius, weatherSystem);
-            }
-            } catch (e) {
-                console.warn(`[World] Failed to spawn procedural extra at ${x},${z}`, e);
-            }
-        };
-
-        // Heuristic to decide if critical based on the random roll
-        // 0.00 - 0.30: Flowers (Deferred)
-        // 0.30 - 0.45: Mushroom (Critical - bouncy)
-        // 0.45 - 0.55: Trees (Critical - obstacle)
-        // 0.55 - 0.75: Musical interactables (Critical)
-        // 0.75 - 0.90: Clouds (Critical if walkable, let's treat all clouds as critical for simplicity)
-        // > 0.90: Spirits, Mirrors, Shrines (Critical)
-
-        const isCritical = rand >= 0.30; // Flowers are < 0.30
-
-        if (isCritical) {
-            spawnExtra();
-            criticalCount++;
-            // Time-based yield after each critical spawn.
-            // Count-based yielding (e.g. every 25 iterations) is unreliable when entity
-            // creation takes 10–30 ms each: 25 × 30 ms = 750 ms worst case per chunk.
-            // Instead we yield as soon as the ENTITY_BUDGET_MS wall-clock budget is
-            // exceeded, keeping every chunk well under 100 ms.
-            if (performance.now() - chunkStart >= ENTITY_BUDGET_MS) {
-                await yieldControl();
-                chunkStart = performance.now();
-            }
-        } else {
-            globalBackgroundProcessor.enqueue({
-                id: `procedural_deferred_${i}`,
-                execute: spawnExtra
-            });
-            deferredCount++;
-        }
-    }
-
-    console.log(`[World] Procedural Extras: ${criticalCount} critical spawned, ${deferredCount} deferred.`);
-}
-
-
-export function spawnNearbyFoliage(origin: THREE.Vector3, type: string, options: FoliageGrowthOptions, weatherSystem: WeatherSystem | null = null): void {
-    if (animatedFoliage.length > 3000) return; // Hard cap
-
-    const maxAttempts = 5;
-    for (let i = 0; i < options.maxOffspring; i++) {
-        if (Math.random() > options.spawnChanceBase) continue;
-
-        let valid = false;
-        let nx = 0, nz = 0;
-
-        for (let a = 0; a < maxAttempts; a++) {
-            const angle = Math.random() * Math.PI * 2;
-            const dist = (Math.random() * 0.5 + 0.5) * options.spawnRadius;
-            nx = origin.x + Math.cos(angle) * dist;
-            nz = origin.z + Math.sin(angle) * dist;
-
-            if (isPositionValid(nx, nz, 1.0)) {
-                // Check local density
-                let localCount = 0;
-                // Use a simple distance check against a subset or the WASM grid if available
-                for (const plant of cpuAnimatedFoliage) {
-                    if (!plant || !plant.position) continue;
-                    const dx = plant.position.x - nx;
-                    const dz = plant.position.z - nz;
-                    if (dx*dx + dz*dz < options.spawnRadius * options.spawnRadius) {
-                        localCount++;
-                    }
-                }
-
-                if (localCount < options.densityLimit) {
-                    valid = true;
-                    break;
-                }
-            }
-        }
-
-        if (valid) {
-            const groundY = getUnifiedGroundHeight(nx, nz);
-            let obj: THREE.Object3D | null = null;
-
-            if (type === 'flower') {
-                obj = createFlower();
-            } else if (type === 'mushroom') {
-                obj = createMushroom({ size: 'regular', scale: 0.8 });
-            }
-
-            if (obj) {
-                obj.position.set(nx, groundY, nz);
-                obj.userData.age = 0;
-                obj.userData.lastSpawnTime = Date.now();
-                safeAddFoliage(obj, false, 0.5, weatherSystem);
-
-                // If it's a batcher-registered object, it will be added to the batcher.
-                // However, safeAddFoliage might push it to arrays that get batched.
-                // The weather ecosystem will handle registering mushrooms.
-            }
-        }
-    }
-}
-
-// Compatibility wrappers for refactored startup flow
-export async function initCriticalWorld(scene: THREE.Scene, weatherSystem?: WeatherSystem): Promise<WorldObjects> {
-    if (!weatherSystem) throw new Error('[World] initCriticalWorld: weatherSystem is required');
-    return initWorld(scene, weatherSystem, false);
-}
-
-export async function initWorldCritical(scene: THREE.Scene, weatherSystem?: WeatherSystem): Promise<WorldObjects> {
-    if (!weatherSystem) throw new Error('[World] initWorldCritical: weatherSystem is required');
-    return initWorld(scene, weatherSystem, false);
-}
-
-export async function initDeferredWorldContent(
-    scene: THREE.Scene,
-    weatherSystem: WeatherSystem,
-    onProgress?: (percent: number, label: string) => void
-): Promise<void> {
-    // Background deferred loading - map generation is triggered separately on enter
-    if (onProgress) onProgress(100, 'Deferred content ready');
-}
-
-export function initWorldContent(scene: THREE.Scene, weatherSystem: WeatherSystem): Promise<void> {
-    return initDeferredWorldContent(scene, weatherSystem);
-}
+export * from './generation-utils.ts';
+export * from './generation-decorators.ts';
+export * from './generation-core.ts';
